+++ src/app.module.ts
@@ -1,8 +1,10 @@
 import { Module } from '@nestjs/common';
+import { APP_FILTER } from '@nestjs/core';
 import { ConfigModule } from '@nestjs/config';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { validateEnv } from './common/config/env.validation';
+import { HttpExceptionFilter } from './common/filters/http-exception.filter';
 import { BankIdModule } from './modules/bankid/bankid.module';
 
 @Module({
@@ -16,6 +18,12 @@ import { BankIdModule } from './modules/bankid/bankid.module';
     BankIdModule,
   ],
   controllers: [AppController],
-  providers: [AppService],
+  providers: [
+    AppService,
+    {
+      provide: APP_FILTER,
+      useClass: HttpExceptionFilter,
+    },
+  ],
 })
 export class AppModule {}